DataProvider fix
